[FEAT] 로또 결과 구현

+++ feature/home/src/main/java/com/hanbang/home/main/navigation/HomeNavigation.kt
@@ -23,7 +23,7 @@ fun NavGraphBuilder.homeNavGraph(
 	onShowErrorSnackBar: (HbSnackBarType) -> Unit,
 	onClickRecommend: () -> Unit,
 	onClickViewMore: () -> Unit,
-	onClickCheckResult: () -> Unit
+	onClickCheckResult: (Int) -> Unit
 ) {
 	composable<RouteHome> {
 		HomeRoute(
